Move JSContext wrapper to script_bindings.
